Enemy can deal damage every 2sec, projectiles deals damage and enemy eventually dies when reaching 0HP or less.

+++ Assets/Scripts/Interfaces/IEnemy.cs
@@ -1,3 +1,3 @@
 ﻿public interface IEnemy {
-    void TakeDamage(int amount);
+    void TakeDamage(float amount);
 }
